+++ src/app/modules/sagas/index.js
@@ -1,5 +1,5 @@
-import modules from '../../modules';
-import resources from '../../resources';
+import modules from '../../experimental-cmf/modules';
+import resources from '../../experimental-cmf/resources';
 
 const resourceConfig = resources({
 	id: 'sagas',
