+++ src/app/selectors/isWebapp.js
@@ -1,4 +1,4 @@
-import modules from '../modules';
+import modules from '../experimental-cmf/modules';
 
 export default function isWebapp(state) {
 	const componentsModule = modules.get('designer.components').inSelector();
